plot_channel_thresholds.py [ok]
