+++ autoreject/utils.py
@@ -4,9 +4,36 @@
 
 import numpy as np
 import mne
+from mne.utils import logger
+
 from sklearn.externals.joblib import Memory
 
 mem = Memory(cachedir='cachedir')
+
+
+def _check_data(epochs, picks, verbose='progressbar'):
+    BaseEpochs = _get_epochs_type()
+    if not isinstance(epochs, BaseEpochs):
+        raise ValueError('Only accepts MNE epochs objects.')
+
+    if epochs.preload is False:
+        raise ValueError('Data must be preloaded.')
+    n_bads = len(epochs.info['bads'])
+
+    picked_info = mne.io.pick.pick_info(epochs.info, picks)
+    ch_types_picked = {
+        mne.io.meas_info.channel_type(picked_info, idx)
+        for idx in range(len(picks))}
+
+    if sum(ch in ch_types_picked for ch in ('mag', 'grad', 'eeg')) > 1:
+        raise ValueError('AutoReject handles only one channel type for now')
+
+    if n_bads > 0:
+        if verbose is not False:
+            logger.info(
+                '%i channels are marked as bad. These will be ignored.'
+                'If you want them to be considered by autoreject please '
+                'remove them from epochs.info["bads"].' % n_bads)
 
 
 def _handle_picks(info, picks):
